refactor: move to more appropriate package
