+++ runtime/src/main/java/io/quarkiverse/power/runtime/PowerSensorProducer.java
@@ -3,6 +3,8 @@ package io.quarkiverse.power.runtime;
 import jakarta.enterprise.inject.Produces;
 import jakarta.inject.Singleton;
 
+import io.quarkiverse.power.runtime.sensors.macos.powermetrics.MacOSPowermetricsSensor;
+
 @Singleton
 public class PowerSensorProducer {
     @Produces
